providers/brandkit_recraft/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
numpy==2.1.1
lxml==5.3.0
